+++ Reinvest/LegalEntities/src/Adapter/Database/DatabaseAdapter.ts
@@ -1,6 +1,7 @@
 import {DatabaseProvider, PostgreSQLConfig} from "PostgreSQL/DatabaseProvider";
 import {Kysely} from "kysely";
 import {
+    LegalEntitiesCompanyAccount,
     LegalEntitiesDraftAccount, LegalEntitiesIndividualAccount,
     LegalEntitiesProfile
 } from "LegalEntities/Adapter/Database/LegalEntitiesSchema";
@@ -8,11 +9,13 @@ import {
 export const legalEntitiesProfileTable = "legal_entities_profile";
 export const legalEntitiesDraftAccountTable = "legal_entities_draft_accounts";
 export const legalEntitiesIndividualAccountTable = "legal_entities_individual_account";
+export const legalEntitiesCompanyAccountTable = "legal_entities_company_account";
 
 export interface LegalEntitiesDatabase {
     [legalEntitiesDraftAccountTable]: LegalEntitiesDraftAccount,
     [legalEntitiesProfileTable]: LegalEntitiesProfile,
     [legalEntitiesIndividualAccountTable]: LegalEntitiesIndividualAccount,
+    [legalEntitiesCompanyAccountTable]: LegalEntitiesCompanyAccount,
 }
 
 export const LegalEntitiesDatabaseAdapterInstanceProvider = "LegalEntitiesDatabaseAdapterProvider";