+++ Reinvest/Registration/src/Domain/EmailCreator.ts
@@ -1,4 +1,4 @@
-import {MappedType} from "Registration/Common/Model/Mapping/MappedType";
+import {MappedType} from "Registration/Domain/Model/Mapping/MappedType";
 
 export class EmailCreator {
     static getClassName = (): string => "EmailCreator";
@@ -8,9 +8,16 @@ export class EmailCreator {
         this.emailDomain = emailDomain;
     }
 
-    create(profileId: string, externalId: string, mappedType: MappedType): string {
+    create(profileId: string, externalId: string, dependentId: string, mappedType: MappedType): string {
         const type = mappedType.toLowerCase();
 
-        return `${type}-${profileId}-${externalId}@${this.emailDomain}`;
+        switch (mappedType) {
+            case MappedType.PROFILE:
+                return `${type}-${profileId}@${this.emailDomain}`;
+            case MappedType.STAKEHOLDER:
+                return `${type}-${profileId}-${externalId}-${dependentId}@${this.emailDomain}`;
+            default:
+                return `${type}-${profileId}-${externalId}@${this.emailDomain}`;
+        }
     }
 }