+++ Reinvest/Registration/src/IntegrationLogic/UseCase/SynchronizeIndividualAccount.ts
@@ -26,9 +26,9 @@ export class SynchronizeIndividualAccount extends AbstractSynchronize {
         this.vertaloSynchronizer = vertaloSynchronizer;
     }
 
-    async execute(record: MappedRecord): Promise<void> {
+    async execute(record: MappedRecord): Promise<boolean> {
         if (!record.isIndividualAccount() || !await this.lockExecution(record)) {
-            return;
+            return false;
         }
         try {
             console.log(`[START] Individual account synchronization, recordId: ${record.getRecordId()}`);
@@ -40,14 +40,16 @@ export class SynchronizeIndividualAccount extends AbstractSynchronize {
             if (northCapitalStatus && vertaloStatus) {
                 await this.setCleanAndUnlockExecution(record);
                 console.log(`[SUCCESS] Individual account synchronized, recordId: ${record.getRecordId()}`);
+                return true;
             } else {
                 console.error(`[FAILED] Individual account synchronization FAILED, recordId: ${record.getRecordId()}`);
                 await this.unlockExecution(record);
             }
         } catch (error: any) {
-            console.error(`[FAILED] Individual account synchronization FAILED with error, recordId: ${record.getRecordId()}: ${error.message}`);
+            console.error(`[FAILED] Individual account synchronization FAILED with error, recordId: ${record.getRecordId()}`, error);
             await this.unlockExecution(record);
         }
+        return false;
     }
 
     private async synchronizeNorthCapital(record: MappedRecord, individualAccount: IndividualAccountForSynchronization): Promise<boolean> {
@@ -59,7 +61,7 @@ export class SynchronizeIndividualAccount extends AbstractSynchronize {
             console.log(`[North Capital SUCCESS] Individual account synchronized, recordId: ${record.getRecordId()}`);
             return true;
         } catch (error: any) {
-            console.error(`[North Capital FAILED] North Capital Individual account synchronization FAILED, recordId: ${record.getRecordId()}: ${error.message}`);
+            console.error(`[North Capital FAILED] North Capital Individual account synchronization FAILED, recordId: ${record.getRecordId()}`, error);
             return false;
         }
     }
@@ -67,12 +69,12 @@ export class SynchronizeIndividualAccount extends AbstractSynchronize {
     private async synchronizeVertalo(record: MappedRecord, individualAccount: IndividualAccountForSynchronization): Promise<boolean> {
         try {
             const vertaloIndividualAccount = VertaloMapper.mapIndividualAccount(individualAccount, record.getEmail());
-            await this.vertaloSynchronizer.synchronizeIndividualAccount(record.getRecordId(), vertaloIndividualAccount)
+            await this.vertaloSynchronizer.synchronizeAccount(record.getRecordId(), vertaloIndividualAccount)
 
             console.log(`[Vertalo SUCCESS] Individual account synchronized, recordId: ${record.getRecordId()}`);
             return true;
         } catch (error: any) {
-            console.error(`[Vertalo FAILED] Vertalo Individual account synchronization FAILED, recordId: ${record.getRecordId()}: ${error.message}`);
+            console.error(`[Vertalo FAILED] Vertalo Individual account synchronization FAILED, recordId: ${record.getRecordId()}`, error);
             return false;
         }
     }