+++ Reinvest/Registration/src/Providers/IntegrationServiceProvider.ts
@@ -6,6 +6,9 @@ import {LegalEntitiesService} from "Registration/Adapter/Modules/LegalEntitiesSe
 import {NorthCapitalSynchronizer} from "Registration/Adapter/NorthCapital/NorthCapitalSynchronizer";
 import {SynchronizeIndividualAccount} from "Registration/IntegrationLogic/UseCase/SynchronizeIndividualAccount";
 import {VertaloSynchronizer} from "Registration/Adapter/Vertalo/VertaloSynchronizer";
+import {SynchronizeCompanyAccount} from "Registration/IntegrationLogic/UseCase/SynchronizeCompanyAccount";
+import {SynchronizeCompany} from "Registration/IntegrationLogic/UseCase/SynchronizeCompany";
+import {SynchronizeStakeholder} from "Registration/IntegrationLogic/UseCase/SynchronizeStakeholder";
 
 
 export class IntegrationServiceProvider {
@@ -19,7 +22,9 @@ export class IntegrationServiceProvider {
         container
             .addSingleton(SynchronizeProfile, [MappingRegistryRepository, LegalEntitiesService, NorthCapitalSynchronizer])
             .addSingleton(SynchronizeIndividualAccount, [MappingRegistryRepository, LegalEntitiesService, NorthCapitalSynchronizer, VertaloSynchronizer])
+            .addSingleton(SynchronizeCompanyAccount, [MappingRegistryRepository, LegalEntitiesService, NorthCapitalSynchronizer, VertaloSynchronizer])
+            .addSingleton(SynchronizeCompany, [MappingRegistryRepository, LegalEntitiesService, NorthCapitalSynchronizer])
+            .addSingleton(SynchronizeStakeholder, [MappingRegistryRepository, LegalEntitiesService, NorthCapitalSynchronizer])
         ;
-
     }
 }